test(go): split report publication tests by platform
Keep portable argument tests available on every host while moving the unchanged owner-only and symlink assertions alongside the Linux secure publisher. Add an explicit non-Linux fail-closed test that requires the unsupported error and no output creation.

Tests: go test -count=1 ./...; go test -race -count=1 ./cmd/ardur-agent-recognition-benchmark; go vet ./...; Linux container execution; scripts/check-local.sh --quick; govulncheck; sealed diff security scan.

Checkpoint: architect/sessions/issue-337-nonlinux-report-test/2026-07-17-journal.md
